nom_l2_switch_controller: implementing an initial version of NOM wrappers.
This is a sketch of the topology.registerWrapper() paradigm. Clients
define their own NOM entities, and then tell pox.topology when and how
to  instantiate those NOM entities.

It's still an open question what the right API is for storing application
specific state in the NOM. This design is not intended to be the final
implementation.