recoco: Fix Task when used with target
The target for a Task has to be a generator for it to be
useful.  Now the default run() wraps it.  It's possible
that it might work to just set .run = target, but this
is more obvious.

--HG--
extra : transplant_source : %AA%04%F6%B4%08Y%A9%A4%D1%04%D1%5B%95%FEY%F4%CCm%C1%60