discovery: New options
There are now two new options.

The first tells discovery to install a flow to send its
packets to the controller.

The second tells it to explicitly drop (send back a null
output) the buffer if it is given a buffer_id.

--HG--
extra : transplant_source : 4%FEdz%8EjkOd%8C%5C%1Ay%04%1A8%03%14%9E%CF