openflow: Arbiter renamed and improved
The arbiter event now gets the connection object
and not just the dpid.  This means, for example,
that one could look at the IPs/ports involved or
the switch features to help decide which nexus
to assign the connection to.

--HG--
extra : transplant_source : %E5%03%09%9B%02%E4%C32%08%3C%7D7%04%16%29no%25%D3%25